Move module files to src/inSituML
